vtpm-exec: Add tool to create vTPM proxy device and launch program
Implement a vtpm-exec tool to create a vTPM proxy device and pass the
necessary parameters to a program it can launch through environment
variables. It also keeps the file descriptor it receives from the
ioctl on /dev/vtpmx open so that the receiving program (shell script)
can use this tool with the --connect-to-ima-ns <fd>  option to connect
the vTPM proxy device to an IMA namespace.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>