UML: Support User Mode Linux (UML) for certain test cases
The user has to set the environment variable IMA_TEST_UML to the full path
of the UML linux binary for UML-supporting test cases to run in UML.

UML test cases are setting up a chroot environment where only busybox is
available as well as the script files relevant for a test case. This
resembles the environment that is setup for container testing. However,
since devices such as /dev/null have to be created it is necessary to
run the UML tests as root.

Test cases for UML need to be converted to not use 'exit' but 'exit_test'
so that in a UML environment the test case can write its exit code into a
file (/__exitcode). The caller then reads the exit code from this file.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>