UML: Run a few test cases entirely in UML using a UML wrapper script
Implement a wrapper script run-in-uml.sh that copies a test case and the
helper script files (common.sh, ns-common.sh, etc.) and executables
(busybox, bash, file, etc.) into a filesystem and starts UML with it and
then runs the typical test.sh file for a test case. This now allows many
of the measurement test cases to entirely run in a UML environment and
also allows to run those test cases that start many containers. However,
test cases that run many concurrent containers are very slow and only
very few containers are created. The latter is due to the fact that under
UML /proc/cpuinfo only shows 1 CPU, which is resolved through our own
nproc tool that always shows 10 CPUs.

Modify imatest to accept a wrapper script (run-in-uml.sh) to run a
test case.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>